feat: update user test case
